--- src/features/paper-list/index.ts
@@ -1,2 +1,0 @@
-export { default as PopularPaper } from "./ui/popular-paper"
-export { default as RecentPaper } from "./ui/recent-paper"
