+++ src/pages/index.ts
@@ -1,2 +1,2 @@
 export { landingRoute } from "./landing/landing-route"
-export { paperListRoute } from "./paper-list/paper-list-route"
+export { paperListRoute } from "./paper-list/ui/paper-list.route"
