+++ src/pages/paper-list/model/use-recent-papers.query.ts
@@ -2,7 +2,7 @@ import { useSuspenseQuery } from "@tanstack/react-query"
 
 import { paperQueryKey, PAPER_LIMIT, PAPER_OFFSET } from "@/entities/paper"
 
-import { getPapers } from "../apis/get-papers"
+import { getPapers } from "../api/get-papers"
 
 import type { PaperParams } from "@/entities/paper"
