+++ src/pages/paper-list/ui/paper-item.tsx
