+++ src/pages/paper-list/ui/paper-list.route.tsx
@@ -1,4 +1,4 @@
-import PaperListPage from "./paper-list-page"
+import PaperListPage from "./paper-list.page"
 
 import type { RouteObject } from "react-router"
