+++ src/pages/paper-list/ui/paper-list.skeleton.tsx
