+++ src/pages/paper-list/ui/paper-list.tsx
