+++ src/pages/paper-list/ui/paper-section.tsx
